requirements/base.txt
litestar[standard]
pydantic-settings
pydantic
uvicorn
sqlalchemy[asyncpg]
advanced-alchemy
asyncpg
bcrypt
pyjwt